requirements.txt
git+https://github.com/openai/swarm.git
python-dotenv==1.0.1
openai==1.53.1